+++ src/main/java/org/folio/qm/util/MarcRecordModifier.java
@@ -4,6 +4,13 @@ import io.vertx.core.json.JsonObject;
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.Optional;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+import java.util.stream.Collectors;
 import lombok.extern.log4j.Log4j2;
 import org.marc4j.MarcJsonReader;
 import org.marc4j.MarcJsonWriter;
@@ -11,6 +18,7 @@ import org.marc4j.MarcStreamWriter;
 import org.marc4j.marc.DataField;
 import org.marc4j.marc.MarcFactory;
 import org.marc4j.marc.Record;
+import org.marc4j.marc.Subfield;
 import org.marc4j.marc.VariableField;
 
 /**
@@ -22,9 +30,15 @@ public final class MarcRecordModifier {
 
   public static final String TAG_999 = "999";
   public static final String TAG_001 = "001";
+  public static final String TAG_003 = "003";
+  public static final String TAG_004 = "004";
+  public static final String TAG_005 = "005";
+  public static final String TAG_035 = "035";
   public static final char INDICATOR_F = 'f';
   public static final char SUBFIELD_I = 'i';
-  public static final char SUBFIELD_S = 's';
+  private static final String OCLC_PREFIX = "(OCoLC)";
+  private static final String OCLC = "OCoLC";
+  private static final String OCLC_PATTERN = "\\((" + OCLC + ")\\)((ocm|ocn|on)?0*|([a-zA-Z]+)0*)(\\d+\\w*)";
 
   private static final MarcFactory FACTORY = MarcFactory.newInstance();
 
@@ -83,6 +97,44 @@ public final class MarcRecordModifier {
       log.error("add001Field:: Failed to add 001 field", e);
       return false;
     }
+  }
+
+  /**
+   * Removes all 003 fields from the MARC record.
+   *
+   * @param marcRecord the marc4j Record to modify
+   */
+  public static void remove003Field(Record marcRecord) {
+    marcRecord.getVariableFields(TAG_003).forEach(marcRecord::removeVariableField);
+  }
+
+  /**
+   * Normalizes 035 fields containing OCLC numbers in the MARC record.
+   * Formats OCLC numbers and removes duplicates.
+   *
+   * @param marcRecord the marc4j Record to modify
+   */
+  public static void normalize035Field(Record marcRecord) {
+    var subfields = get035SubfieldOclcValues(marcRecord, TAG_035);
+    if (!subfields.isEmpty()) {
+      formatOclc(subfields);
+      deduplicateOclc(marcRecord, subfields, TAG_035);
+    }
+  }
+
+  /**
+   * Retrieves the data from control field 004.
+   *
+   * @param marcRecord the marc4j Record
+   * @return the data of control field 004, or null if not present
+   */
+  public static String get004ControlFieldData(Record marcRecord) {
+    return marcRecord.getControlFields()
+      .stream()
+      .filter(controlField -> controlField.getTag().equals(TAG_004))
+      .findFirst()
+      .map(controlField -> controlField.getData().trim())
+      .orElse(null);
   }
 
   /**
@@ -148,5 +200,80 @@ public final class MarcRecordModifier {
       }
     }
     return null;
+  }
+
+  private static List<Subfield> get035SubfieldOclcValues(Record srcRecord, String tag) {
+    return Optional.ofNullable(srcRecord)
+      .stream()
+      .flatMap(marcRecord -> marcRecord.getVariableFields(tag).stream())
+      .flatMap(field -> get035oclcSubfields(field).stream())
+      .collect(Collectors.toList());
+  }
+
+  private static List<Subfield> get035oclcSubfields(VariableField field) {
+    if (field instanceof DataField dataField) {
+      return dataField.getSubfields().stream()
+        .filter(sf -> sf.getData().trim().startsWith(OCLC_PREFIX))
+        .toList();
+    }
+    return Collections.emptyList();
+  }
+
+  private static void formatOclc(List<Subfield> subfields) {
+    Pattern pattern = Pattern.compile(OCLC_PATTERN);
+
+    for (Subfield subfield : subfields) {
+      String data = subfield.getData().replaceAll("[.\\s]", "");
+      Matcher matcher = pattern.matcher(data);
+      if (matcher.find()) {
+        String oclcTag = matcher.group(1); // "OCoLC"
+        String numericAndTrailing = matcher.group(5); // Numeric part and any characters that follow
+        String prefix = matcher.group(2); // Entire prefix including letters and potentially leading zeros
+
+        if (prefix != null && (prefix.startsWith("ocm") || prefix.startsWith("ocn") || prefix.startsWith("on"))) {
+          // If "ocm" or "ocn", strip entirely from the prefix
+          subfield.setData("(" + oclcTag + ")" + numericAndTrailing);
+        } else {
+          // For other cases, strip leading zeros only from the numeric part
+          numericAndTrailing = numericAndTrailing.replaceFirst("^0+", "");
+          if (prefix != null) {
+            prefix = prefix.replaceAll("\\d+", ""); // Safely remove digits from the prefix if not null
+          }
+          // Add back any other prefix that might have been included like "tfe"
+          subfield.setData("(" + oclcTag + ")" + (prefix != null ? prefix : "") + numericAndTrailing);
+        }
+      }
+    }
+  }
+
+  private static void deduplicateOclc(Record marcRecord, List<Subfield> subfields, String tag) {
+    List<Subfield> subfieldsToDelete = new ArrayList<>();
+
+    for (Subfield subfield : new ArrayList<>(subfields)) {
+      if (subfields.stream().anyMatch(s -> isOclcSubfieldDuplicated(subfield, s))) {
+        subfieldsToDelete.add(subfield);
+        subfields.remove(subfield);
+      }
+    }
+    var variableFields = marcRecord.getVariableFields(tag);
+    subfieldsToDelete.forEach(subfieldToDelete ->
+      variableFields.forEach(field -> removeSubfieldIfExist(marcRecord, field, subfieldToDelete)));
+  }
+
+  private static boolean isOclcSubfieldDuplicated(Subfield s1, Subfield s2) {
+    return !s1.equals(s2)
+      && s1.getData().equals(s2.getData())
+      && s1.getCode() == s2.getCode();
+  }
+
+  private static void removeSubfieldIfExist(Record marcRecord, VariableField field,
+                                            Subfield subfieldToDelete) {
+    if (field instanceof DataField dataField && dataField.getSubfields().contains(subfieldToDelete)) {
+      if (dataField.getSubfields().size() > 1) {
+        dataField.removeSubfield(subfieldToDelete);
+      } else {
+        marcRecord.removeVariableField(dataField);
+      }
+    }
   }
 }