+++ src/test/java/org/folio/qm/service/change/HoldingsChangeRecordServiceTest.java
@@ -15,6 +15,7 @@ import static org.mockito.Mockito.when;
 import io.vertx.core.json.JsonObject;
 import java.util.Collections;
 import java.util.List;
+import java.util.UUID;
 import org.folio.ExternalIdsHolder;
 import org.folio.ParsedRecord;
 import org.folio.RawRecord;
@@ -118,7 +119,7 @@ class HoldingsChangeRecordServiceTest {
     verify(validationService).validateMarcRecord(any(BaseQuickMarcRecord.class),
       eq(Collections.emptyList()));
     verify(validationService).validate(quickMarcEdit);
-    verify(sourceRecordService).update(eq(recordId), any(Record.class));
+    verify(sourceRecordService).update(eq(UUID.fromString(existingSourceRecord.getMatchedId())), any(Record.class));
     verify(folioRecordService).update(quickMarcEdit.getExternalId(), holdingsRecord);
   }
 
@@ -371,6 +372,7 @@ class HoldingsChangeRecordServiceTest {
 
     return new Record()
       .withId(recordId)
+      .withMatchedId(recordId)
       .withRecordType(RecordType.MARC_HOLDING)
       .withGeneration(0)
       .withRawRecord(new RawRecord().withId(recordId))