+++ src/test/java/org/folio/qm/service/change/InstanceChangeRecordServiceTest.java
@@ -15,6 +15,7 @@ import static org.mockito.Mockito.when;
 import io.vertx.core.json.JsonObject;
 import java.util.Collections;
 import java.util.List;
+import java.util.UUID;
 import org.folio.ExternalIdsHolder;
 import org.folio.ParsedRecord;
 import org.folio.RawRecord;
@@ -121,7 +122,7 @@ class InstanceChangeRecordServiceTest {
     verify(defaultValuesPopulationService).populate(quickMarcEdit);
     verify(validationService).validateMarcRecord(any(BaseQuickMarcRecord.class), eq(Collections.emptyList()));
     verify(validationService).validate(quickMarcEdit);
-    verify(sourceRecordService).update(eq(recordId), any(Record.class));
+    verify(sourceRecordService).update(eq(UUID.fromString(existingSourceRecord.getMatchedId())), any(Record.class));
     verify(folioRecordService).update(quickMarcEdit.getExternalId(), instanceRecord);
     verify(linksService).updateRecordLinks(quickMarcRecord);
   }
@@ -376,6 +377,7 @@ class InstanceChangeRecordServiceTest {
 
     return new Record()
       .withId(recordId)
+      .withMatchedId(recordId)
       .withRecordType(RecordType.MARC_BIB)
       .withGeneration(0)
       .withRawRecord(new RawRecord().withId(recordId))